refactor(mypage): change components location
