+++ components/ranking/tier.ts
@@ -1,0 +1,7 @@
+import bronze from "../../public/Tier/Bronze.png";
+import silver from "../../public/Tier/Silver.png";
+import gold from "../../public/Tier/Gold.png";
+import platinum from "../../public/Tier/Platinum.png";
+import diamond from "../../public/Tier/Diamond.png";
+
+export { bronze, silver, gold, platinum, diamond };
